// VueMyWebGIS/baseSystem/src/api/path.ts
export enum Api {
  //
  // GET_LOGIN = '/index/login',
  GET_LOGIN = "login/userlogin",
  GetForgetPasswordApi = 'index/forget_password',
  GET_REGISTER = 'index/register',
  GET_DELETE = "index/delete",
 
  GET_HOTEL_ROOM = 'index/hotel_room',
  GET_ROOM_DETAIL = 'index/room_detail',
  GET_USER_COUNT = 'user/count',

  GET_USER_ActiveCOUNT = 'user/ActiveCOUNT',
  GET_SaleCount = 'Sale/Count',
  GET_AllIncome = 'Sale/AllIncome',
  GET_ReportData = "Report/Data",
  GET_ReportList = "Report/List",
  GET_ReportViewData = "Report/ViewData",
  getReportGMV = "Report/GMV",
  getReportOrderCount = "Report/OrderCount",
  getReportActiviesUserCount = "Report/ActiviesUserCount",
  getReportChangeRace = "Report/ChangeRace",
  getPayrecordData = "Pay/PayrecordData",
  getDataBaseQuery = "Database/Query",
  getHotelRoomList = "Hotel/RoomList",
  getHotelRoomListTop6 = "Hotel/RoomListTop6",
  getCarouselList = "Hotel/CarouselList",
  getTitleList = "Hotel/TitleList",
  getTopLogList = "Hotel/TopLogList",
  getDynamicValue = "Dynamic/DynamicValue",
  getMSGValue = "Dynamic/MSGValue",
  getCollectionList = "Dynamic/CollectionList",
  GetHistoryList = "Dynamic/HistoryList",
  GetMessageList = "Dynamic/MessageList",
  GetDynamicList = "Dynamic/DynamicList",
  getHeaderMenuList = "Header/HeaderMenuList", 
  getHeaderInfo = "Header/HeaderInfo",
  getVehicles = "Vehicles/List",
  getVehiclesTable = "Vehicles/table",
  getVehiclesCardStatus = "Vehicles/CardStatus",
  getVehiclesTrackings = "Vehicles/Trackings",
  getVehiclePointInfo = "Vehicle/VehiclePointInfo",
}